--- a/RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Helpers/AConsoleOutput.cs.meta
+++ b/RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Helpers/AConsoleOutput.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: c3e72249de38b82478f463072e36af13
-timeCreated: 1458773522
+timeCreated: 1460161381
 licenseType: Free
 MonoImporter:
   serializedVersion: 2