… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Helpers/MemberType.cs.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Parser.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Parser/IRexIntellisenceProvider.cs.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Parser/IRexParser.cs.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/Parser/RexParser.cs.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/RexReflectionHelper.cs.meta
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Test/InputStateMachineTest.cs
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Test/IntelliSenceTests.cs
… RexWindowProjcet/Assets/Editor/RexDiagnostics/Test/RexHelperTest.cs
… RexWindowProjcet/Assets/Editor/RexDiagnostics/UI/OutputEntry.cs.meta
… RexWindowProjcet/RexWindowProjcet.CSharp.Editor.csproj