## Back Story

When unity editor enters or exits play mode it reloads all scripts in a following order:

1. Serialize all managed code refernaces.
2. Store the serialized data in c++ space.
3. Unload all managed scripts (C# scripts)
4. Compile C# code,
5. Initialize the managed space.
6. Deserialize scripts from c++ back into managed space.
 
Editor windows need to account for that and store data either in EditorPrefs or have it setup in a way that enables unity to serialize the window data.
Here are two resources I found helpful:
http://blogs.unity3d.com/2014/06/24/serialization-in-unity/
http://forum.unity3d.com/threads/serialization-best-practices-megapost.155352/

Improve startup time
- 

In this PR I reduced the number of places where object had static initialization since that is called in step five and happens on every load. 

### Improvements:
- Lazy loading for RexUtils class instead of loading on each script compile.
- Faster mapping to primitives functions in RexReflectionHelper.
- RexParser now lazy loads the regex and this reduces the load time on script compile.
- Redesigned the output architecture.
- Reduced the use of static field initialization
- More summaries added
- Renames of poorly named classes.

*Before:*
![2016-04-09 02_55_11-postur](https://cloud.githubusercontent.com/assets/2693474/14401364/1b0da660-fdff-11e5-92c0-b14d86f91021.png)

*After*
![2016-04-09 02_54_18-postur](https://cloud.githubusercontent.com/assets/2693474/14401365/1b221672-fdff-11e5-8624-a505fae2af00.png)

- | Before | After | Difference
--- | --- | --- | ---
Startup Time | ~4,6ms   | ~2,5ms | -54%
Memory Alloc | 90,7 KB | 3 KB |  -96%